requirements.txt
os
flask
flask_session
werkzeug.security
sqlite3
functools
datetime